refactor: move react-query fetching logic to page level
The react-query fetching logic was moved from the column components to the page level.
This was done to improve performance and reduce the number of times the data is fetched.
The column components now receive the data as props.
This also removes the need for the Loading component in the column components.