feat: add github workflow to deploy the app to digital ocean
This commit introduces a new GitHub Actions workflow
defined in `.github/workflows/deploy.yml`. This workflow
automates the deployment process to a DigitalOcean server
whenever changes are pushed to the `main` branch.

The workflow includes steps for:
- Checking out the code
- Setting up pnpm and Node.js
- Installing dependencies
- Generating OG images
- Building the application
- Deploying the built files to the DigitalOcean server
- Restarting the Nginx server to serve the updated files.

Secrets are used to securely provide the necessary
credentials for accessing the DigitalOcean server.